+++ data-raw/agreements/GNEVAR/prepare-GNEVAR.R
@@ -10,13 +10,18 @@ GNEVAR <- readr::read_csv("data-raw/agreements/GNEVAR/EnvGov Nodes-Table 1 VERS2
 # In this stage you will want to correct the variable names and
 # formats of the 'GNEVAR' object until the object created
 # below (in stage three) passes all the tests. 
-GNEVAR <- as_tibble(GNEVAR)  %>% 
-  transmutate(Beg = standardise_dates(DocSign),
+GNEVAR <- as_tibble(GNEVAR)  %>%
+  tidyr::separate(IEA, c("NEW", "IEADB_ID"), sep = "-") %>% 
+  dplyr::mutate(D=dplyr::recode(T, G="A", M="E", "T"="Q", D="V", R="W", N="X", U="Y")) %>% 
+  transmutate(Signature = standardise_dates(DocSign),
               End = standardise_dates(DocEnd),
               Force = standardise_dates(DocForce),# some dates formats are failing to pass (e.i 0000-00-00)
-              ID = GENG) %>%
-  dplyr::select(ID, Title, Beg, End) %>% 
-  dplyr::arrange(Beg, ID)
+              GNEVAR_ID = GENG,
+              ECOLEX_ID = ECOLEX) %>% 
+  dplyr::mutate(Title = standardise_titles(Title)) %>%
+  dplyr::mutate(Beg = dplyr::coalesce(Signature, Force)) %>% 
+  dplyr::select(GNEVAR_ID, Title, Beg, End, L,J,D, Signature, Force, IEADB_ID, ECOLEX_ID) %>% 
+  dplyr::arrange(Beg, GNEVAR_ID)
 
 # qData includes several functions that should help cleaning and standardising your data.
 # Please see the vignettes or website for more details.